feat : category page - CategoryBar component
Related to : #99
