+++ src/styles/common/categoryBox.module.css
@@ -36,7 +36,6 @@
 }
 
 /* 카테고리 박스 하단 줄 */
-
 .categoryBottom {
 	width: inherit;
 	height: 1px;
